+++ rules/rule_register.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/Azure/tflint-ruleset-avm/interfaces"
 	"github.com/Azure/tflint-ruleset-avm/outputs"
-	"github.com/Azure/tflint-ruleset-avm/requiredversion"
 	"github.com/Azure/tflint-ruleset-avm/waf"
 	azurerm "github.com/Azure/tflint-ruleset-azurerm-ext/rules"
 	basic "github.com/Azure/tflint-ruleset-basic-ext/rules"
@@ -24,11 +23,11 @@ var Rules = func() []tflint.Rule {
 			Wrap(basic.NewTerraformVariableSeparateRule()),
 			Wrap(azurerm.NewAzurermResourceTagRule()),
 			NewTerraformDotTfRule(),
+			NewRequiredVersionRule(),
 		},
 		waf.Rules,
 		interfaces.Rules,
 		outputs.Rules,
-		requiredversion.Rules,
 	)
 }()
